+++ sdk/include/reactos/CtfImeTable.h
