+++ sdk/include/reactos/imetable.h
